refactor(tests): move ensure_shared_memory_client to test_support module
- extract implementation from mod.rs into a dedicated test_support.rs
  so mod.rs stays export-focused (per coding guidelines)
- change `let _ = global::init(...)` to `.expect(...)` so init failures
  surface as clear test panics rather than silent swallowed errors

addresses CodeRabbit suggestions on PR #2646
